[#107] refactor: seperate `ChordInSegment` to swift file
